change Entity, add changelog for Liquibase

+++ src/main/java/com/edu/ulab/app/service/impl/BookServiceImpl.java
@@ -2,16 +2,20 @@ package com.edu.ulab.app.service.impl;
 
 import com.edu.ulab.app.dto.BookDto;
 import com.edu.ulab.app.entity.Book;
+import com.edu.ulab.app.entity.Person;
 import com.edu.ulab.app.exception.BookNotFoundException;
 import com.edu.ulab.app.mapper.dto.BookDtoMapper;
 import com.edu.ulab.app.repository.BookRepository;
 import com.edu.ulab.app.service.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.hibernate.mapping.Collection;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -66,7 +70,7 @@ public class BookServiceImpl implements BookService {
     @Override
     public List<BookDto> getBooksByUserId(Long id) {
         log.info("Get user books by userID: {}", id);
-        List<Book> books = bookRepository.findBooksByUserId(id);
+        List<Book> books = bookRepository.findBooksByUserId(id.intValue());
         log.info("A books list was found: {}", books);
         return books.stream().map(bookDtoMapper::bookToBookDto).toList() ;
     }
@@ -74,9 +78,10 @@ public class BookServiceImpl implements BookService {
     @Override
     public List<Long> getBooksIdByUserId(Long id) {
         log.info("Get user booksID by userID: {}", id);
-        List<Book> books = bookRepository.findBooksByUserId(id);
+        List<Book> books = bookRepository.findBooksByUserId(id.intValue());
         log.info("A books list was found: {}", books);
-        return books.stream().map(Book::getId).toList() ;
+
+        return books.stream().map(Book::getId).toList();
     }
